solved the logs registring problem
question was: how to register thread logs to logger

what we do :

Logger contains an Arc<Mutex<LinkedList<Storage<...>>>>

Logger is created first.
then its pool_builder method will create a pool builder
which will receive a copy of the Arc
builder will forward info to the Registry.
when a thread is created it will append its storage to the list.

difficulty was :
- what if thread does not start immediately ?
- what if Logger is dropped before pool ?
- how to choose which threads to log ?....
....

this solution seems nice